+++ openNPL/urls.py
@@ -37,7 +37,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include, re_path
 
-from . import views, settings
+from . import npl_views, sflp_views, settings
 
 from rest_framework import permissions
 from drf_yasg.views import get_schema_view
@@ -47,7 +47,7 @@ schema_view = get_schema_view(
    openapi.Info(
       title="openNPL API",
       default_version='v1',
-      description="openNPL is an open source platform for the management of non-performing loans. It implements detailed European Banking Authority loan templates for NPL data",
+      description="openNPL is an open source platform for the management of credit risk and non-performing loans. It implements detailed European Banking Authority loan templates for NPL data and US Agency Single Family Loan Performance Data",
       terms_of_service="https://www.openriskmanagement.com/",
       contact=openapi.Contact(email="info@openrisk.eu"),
       license=openapi.License(name="MIT License"),
@@ -58,11 +58,12 @@ schema_view = get_schema_view(
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('grappelli/', include('grappelli.urls')),  # grappelli URLS
-    path('grappelli/grp_doc/', include('grappelli.urls_docs')),  # grappelli URLS
+    # path('grappelli/', include('grappelli.urls')),  # grappelli URLS
+    # path('grappelli/grp_doc/', include('grappelli.urls_docs')),  # grappelli URLS
     path('', include('start.urls')),  # start URLS
-    path(r'api/', views.api_root, name='api_root'),  # API root
+    path(r'api/', npl_views.api_root, name='api_root'),  # Overall API root
     path(r'api/npl_data/', include(('npl_portfolio.urls', 'npl_portfolio'), namespace='npl_portfolio')),  # NPL data
+    path(r'api/sflp_data/', include(('sflp_portfolio.urls', 'sflp_portfolio'), namespace='sflp_portfolio')),  # SFLP data
     re_path(r'^api/swagger(?P<format>\.json|\.yaml)$', schema_view.without_ui(cache_timeout=0), name='schema-json'),
     re_path(r'^api/swagger/$', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
     re_path(r'^api/redoc/$', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),