+++ sflp_portfolio/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class EbaPortfolioConfig(AppConfig):
+    name = 'sflp_portfolio'