+++ sflp_portfolio/property_collateral.py
@@ -1,0 +1,66 @@
+# Copyright (c) 2020 - 2022 Open Risk (https://www.openriskmanagement.com)
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+
+from django.db import models
+from django.urls import reverse
+
+from sflp_portfolio.property_collateral_choices import *
+from sflp_portfolio.loan import Loan
+
+
+class PropertyCollateral(models.Model):
+    """
+    The PropertyCollateral model object holds Property Collateral data
+
+    .. note:: The Agency Single Family Loan Performance Template does not explicitly segment data attributes into Counterparty, Loan etc. The assignment into tables (models) in openNPL is based on the interpretation and main function of different data fields
+
+    """
+
+    #
+    # IDENTIFICATION FIELDS
+    #
+
+
+    #
+    # FOREIGN KEYS
+    #
+
+    loan_identifier = models.ForeignKey(Loan, on_delete=models.CASCADE, null=True, blank=True)
+
+    #
+    # DATA PROPERTIES
+    #
+
+
+
+    # Bookkeeping fields
+    creation_date = models.DateTimeField(auto_now_add=True)
+    last_change_date = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return self.protection_identifier
+
+    def get_absolute_url(self):
+        return reverse('sflp_portfolio:property_collateral_edit', kwargs={'pk': self.pk})
+
+    class Meta:
+        verbose_name = "Property Collateral"
+        verbose_name_plural = "Property Collateral"